+++ cmd/gatecheck/main.go
@@ -2,13 +2,16 @@ package main
 
 import (
 	"errors"
-	"github.com/gatecheckdev/gatecheck/cmd"
-	"github.com/gatecheckdev/gatecheck/pkg/epss"
-	"github.com/gatecheckdev/gatecheck/pkg/export/defectdojo"
 	"net/http"
 	"os"
 	"strings"
 	"time"
+
+	"github.com/gatecheckdev/gatecheck/cmd"
+	"github.com/gatecheckdev/gatecheck/internal/log"
+	"github.com/gatecheckdev/gatecheck/pkg/epss"
+	"github.com/gatecheckdev/gatecheck/pkg/export/defectdojo"
+	"github.com/spf13/cobra"
 )
 
 const ExitSystemFail int = -1
@@ -49,8 +52,17 @@ func main() {
 		PipedInput:         pipedFile,
 	})
 
+	command.PersistentPreRun = func(_ *cobra.Command, _ []string) {
+		if cmd.GlobalVerboseOutput == false {
+			log.SetLogLevel(log.Disabled)
+		}
+		log.StartCLIOutput(command.ErrOrStderr())
+	}
+
 	command.SilenceUsage = true
+
 	err := command.Execute()
+	log.Info("**** Command Execution Complete ****")
 
 	if errors.Is(err, cmd.ErrorFileAccess) {
 		command.PrintErrln(err)